LayoutTests/http/tests/site-isolation/pagehide-event-iframe.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<!-- Similar to pagehide-event.html, but tests that event is fired in iframes. -->

<!DOCTYPE html>
<title>Tests that pagehide is fired in iframes when site isolation is enabled.</title>

<script src="/js-test-resources/js-test.js"></script>
<script>
description("Tests that pagehide is fired in iframes when site isolation is enabled.");
jsTestIsAsync = true;

const chan = new BroadcastChannel("pagehide-chan");
chan.addEventListener("message", (event) => {
  shouldBeEqualToString("event.data", "pagehide");
  finishJSTest();
})

</script>

<!-- This iframe is same-origin with this page. -->
<iframe src="/site-isolation/resources/pagehide-event-subpage.html"></iframe>
